requirements.txt
blinker==1.9.0
click==8.3.2
filelock==3.25.2
Flask==3.1.3
fsspec==2026.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
pillow==12.2.0
setuptools==81.0.0
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
typing_extensions==4.15.0
Werkzeug==3.1.8
